+++ locations.py
@@ -1,4 +1,23 @@
 # Josh Lim
 # Comp Sci 30 P4
 # 11/07/2019
-# Dictionary containing all the locations
+# Dictionary and list containing all the locations
+
+
+escapeRoom = {
+    "cabinet": ["go back"],
+    "chest": ["go back"],
+    "bookcase": ["go back"],
+    "painting": ["go back"],
+    "desk": ["go back"],
+    "shelve": ["go back"],
+    "table": ["go back"],
+    "chair": ["go back"],
+    "entertainment center": ["go back"],
+}
+
+cabinet = {
+    "description": "",
+    "key": True,
+    "description": "",
+}