+++ properties.py
@@ -4,7 +4,7 @@
 # Nested dictionaries for game properties
 
 
-def nestedDictPrinter(dictName):
+def printNDict(dictName):
     """Prints out content in a nested dictionary with up to 2 layers,
     takes dictName as the dictionary name.
     """
@@ -18,56 +18,16 @@ def nestedDictPrinter(dictName):
             key along with the value with a tab before.
             """
             print(f"\t{items.title()} - {key[items]}")
-# Nested dictionary containing information for inventory
-inventory = {
-    "stick": {
-        "ATTACK": 3,
-        "DURABILITY": 5,
-        },
-    "iron sword": {
-        "ATTACK": 15,
-        "DURABILITY": 20,
-        "SHARPNESS": 10,
-        }
-    }
-# Nested dictionary containing information for mobs
-mobs = {
-    "spider": {
-        "ATTACK": 3,
-        "HIT POINTS": 10,
-        "VITALITY": 10,
-        "DEFENSE": 10,
-        "LUCK": 0,
-        }
-    }
-# Nested dictionary containing information for character
-character = {
-    "stats": {
-        "ATTACK": 10,
-        "HIT POINTS": 100,
-        "VITALITY": 10,
-        "DEFENSE": 10,
-        "LUCK": 0,
-        }
-}
-# Nested dictionary containing information for locations
-locations = {
-    "path 1": {
-        "description": "You have encountered a level 1 mob!",
-        },
-    "path 2": {
-        "description": "You are greeted with 2 paths, which will you choose?",
-        },
-    "path 3": {
-        "description": "You find yourself in a forest, what will you do?",
-    }
-}
 
 escapeRoom = {
-    "description": "you find yourself in a room, in it you see...",
+    "description": "you find yourself in a room and in it you see...",
     "cabinet": ["go back"],
     "chest": ["go back"],
     "bookcase": ["go back"],
     "painting": ["go back"],
     "desk": ["go back"],
+    "shelve": ["go back"],
+    "table": ["go back"],
+    "chair": ["go back"],
+    "entertainment center": ["go back"],
 }